+++ client/src/components/Admin/AdminSideBar.jsx
@@ -1,0 +1,99 @@
+import * as React from "react";
+import { styled, useTheme } from "@mui/material/styles";
+import Box from "@mui/material/Box";
+import MuiDrawer from "@mui/material/Drawer";
+import List from "@mui/material/List";
+import IconButton from "@mui/material/IconButton";
+import DashboardIcon from "@mui/icons-material/Dashboard";
+import HomeRepairServiceIcon from "@mui/icons-material/HomeRepairService";
+import PeopleAltIcon from "@mui/icons-material/PeopleAlt";
+import FormatListBulletedIcon from "@mui/icons-material/FormatListBulleted";
+import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
+import ChevronRightIcon from "@mui/icons-material/ChevronRight";
+import ListItem from "@mui/material/ListItem";
+import ListItemIcon from "@mui/material/ListItemIcon";
+import ListItemText from "@mui/material/ListItemText";
+
+const drawerWidth = 240;
+
+//Opciones de sideBar. Si se necesita agregar mas botones a la lista
+//solo se agrega a este objeto una nueva propiedad con un componente Material Icon como value
+const sideBarOptions = {
+  Dashboard: <DashboardIcon />,
+  Services: <HomeRepairServiceIcon />,
+  Users: <PeopleAltIcon />,
+  "Categories/Groups": <FormatListBulletedIcon />,
+};
+
+const openedMixin = (theme) => ({
+  width: drawerWidth,
+  transition: theme.transitions.create("width", {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.enteringScreen,
+  }),
+  overflowX: "hidden",
+});
+
+const closedMixin = (theme) => ({
+  transition: theme.transitions.create("width", {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  overflowX: "hidden",
+  width: `calc(${theme.spacing(7)} + 1px)`,
+});
+
+const Drawer = styled(MuiDrawer, {
+  shouldForwardProp: (prop) => prop !== "open",
+})(({ theme, open }) => ({
+  width: drawerWidth,
+  flexShrink: 0,
+  whiteSpace: "nowrap",
+  boxSizing: "border-box",
+  ...(open && {
+    ...openedMixin(theme),
+    "& .MuiDrawer-paper": openedMixin(theme),
+  }),
+  ...(!open && {
+    ...closedMixin(theme),
+    "& .MuiDrawer-paper": closedMixin(theme),
+  }),
+}));
+
+export default function AdminSideBar() {
+  const theme = useTheme();
+  const [open, setOpen] = React.useState(false);
+
+  const handleDrawerOpen = () => {
+    setOpen(true);
+  };
+
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
+
+  return (
+    <Box sx={{ display: "flex" }}>
+      <Drawer variant="permanent" open={open}>
+        <List>
+          <ListItem button key={"openDrawerButton"}>
+            <IconButton
+              onClick={open ? handleDrawerClose : handleDrawerOpen}
+              sx={{ ml: "auto" }}
+            >
+              {open ? <ChevronLeftIcon /> : <ChevronRightIcon />}
+            </IconButton>
+          </ListItem>
+          {Object.keys(sideBarOptions).map((text, index) => (
+            <ListItem button key={text}>
+              <ListItemIcon sx={{ fontSize: "large" }}>
+                {sideBarOptions[text]}
+              </ListItemIcon>
+              <ListItemText primary={text} />
+            </ListItem>
+          ))}
+        </List>
+      </Drawer>
+    </Box>
+  );
+}
